Code/Pseudos setup: Separate frontend from backend
