+++ src/aiidalab_qe/setup/__init__.py
@@ -1,0 +1,3 @@
+"""Package for the QE app CLI for setting up codes and pseudos"""
+
+__all__ = []
